Move logging initialization to androidx.startup
