+++ app/src/main/java/org/jellyfin/androidtv/LogInitializer.kt
@@ -1,0 +1,28 @@
+package org.jellyfin.androidtv
+
+import android.content.Context
+import androidx.startup.Initializer
+import timber.log.Timber
+
+class LogInitializer : Initializer<Unit> {
+	override fun create(context: Context) {
+		// Enable improved logging for leaking resources
+		// https://wh0.github.io/2020/08/12/closeguard.html
+		if (BuildConfig.DEBUG) {
+			try {
+				Class.forName("dalvik.system.CloseGuard")
+					.getMethod("setEnabled", Boolean::class.javaPrimitiveType)
+					.invoke(null, true)
+			} catch (e: ReflectiveOperationException) {
+				@Suppress("TooGenericExceptionThrown")
+				throw RuntimeException(e)
+			}
+		}
+
+		// Initialize the logging library
+		Timber.plant(Timber.DebugTree())
+		Timber.i("Debug tree planted")
+	}
+
+	override fun dependencies() = emptyList<Class<out Initializer<*>>>()
+}
