+++ app/src/main/java/org/jellyfin/androidtv/di/KoinInitializer.kt
@@ -2,6 +2,7 @@ package org.jellyfin.androidtv.di
 
 import android.content.Context
 import androidx.startup.Initializer
+import org.jellyfin.androidtv.LogInitializer
 import org.koin.android.ext.koin.androidContext
 import org.koin.core.KoinApplication
 import org.koin.core.context.startKoin
@@ -21,5 +22,6 @@ class KoinInitializer : Initializer<KoinApplication> {
 		)
 	}
 
-	override fun dependencies() = emptyList<Class<out Initializer<*>>>()
+	override fun dependencies() = listOf(LogInitializer::class.java)
 }
+
